+++ CHANGELOG.md
@@ -1,4 +1,5 @@
-* Finders now managed by guice and any additional aop could be applied; abstract beans could define finder methods.
+### 2.1.0 (2015-01-06)
+* Finders now managed by guice and any additional aop could be applied. Abstract beans could define finder methods.
 (integration implemented in [external library](https://github.com/xvik/guice-ext-annotations))
 * Finders auto scan api deprecated: now finder bean could be marked with @ProvidedBy and rely on guice JIT resolution
 * Fix transactional aop warnings on jdk8